refactored question view page
