+++ studyAi/src/app/library/question/components/client/questionsView.tsx
@@ -1,10 +1,6 @@
 "use client";
-import ContainerBar, { Container } from "./containerBar";
-import capitalizeEveryWord from "@/app/util/parsers/capitalizeEveryWord";
-import EditIcon from "@mui/icons-material/Edit";
-import { Button, Chip, IconButton, Tab, Tabs } from "@mui/material";
-import { useState } from "react";
-import { useSession } from "next-auth/react";
+import { Container } from "../server/containerBar";
+import { Chip, IconButton } from "@mui/material";
 import { useQuestions } from "@/app/stores/questionStore";
 import { useParams } from "next/navigation";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
@@ -12,8 +8,6 @@ import { Share } from "@mui/icons-material";
 import { faThumbsUp, faThumbsDown } from "@fortawesome/free-regular-svg-icons";
 import { faPlus } from "@fortawesome/free-solid-svg-icons";
 import { parseInteger } from "@/app/util/parsers/parseInt";
-const containerTabs = ["description", "solution", "attempts"] as const;
-
 const QuestionActionBtns = () => {
   return (
     <div className="flex items-center space-x-2">
@@ -51,7 +45,7 @@ const LikeCounterBtns = () => {
     </div>
   );
 };
-const InnerContainer = ({ view }: { view: (typeof containerTabs)[number] }) => {
+export const QuestionView = () => {
   const params = useParams();
   const questions = useQuestions()[0].data;
   const question =
@@ -88,57 +82,3 @@ const InnerContainer = ({ view }: { view: (typeof containerTabs)[number] }) => {
     </Container>
   );
 };
-const TopBar = ({
-  view,
-  handleChange,
-}: {
-  view: (typeof containerTabs)[number];
-  handleChange: (
-    event: React.SyntheticEvent,
-    newValue: (typeof containerTabs)[number]
-  ) => void;
-}) => {
-  const params = useParams();
-  const session = useSession();
-  const questions = useQuestions()[0].data;
-  const question =
-    params.id && typeof params.id === "string" ? questions[params.id] : null;
-  return (
-    <ContainerBar>
-      <Tabs value={view} onChange={handleChange} aria-label="question-options">
-        {containerTabs.map((tab) => (
-          <Tab
-            key={tab}
-            value={tab}
-            label={capitalizeEveryWord(tab)}
-            sx={{
-              textTransform: "none",
-            }}
-          />
-        ))}
-      </Tabs>
-      {session.data &&
-        question &&
-        session.data.user.id === question.creatorId && (
-          <Button sx={{ textTransform: "none" }}>
-            <EditIcon />
-          </Button>
-        )}
-    </ContainerBar>
-  );
-};
-export const QuestionContainer = ({ height }: { height?: string | number }) => {
-  const [view, setView] =
-    useState<(typeof containerTabs)[number]>("description");
-  const handleChange = (
-    event: React.SyntheticEvent,
-    newValue: (typeof containerTabs)[number]
-  ) => setView(newValue);
-  return (
-    <Container style={{ height: height ? height + "px" : undefined }}>
-      <TopBar view={view} handleChange={handleChange} />
-      <InnerContainer view={view} />
-    </Container>
-  );
-};
-export default QuestionContainer;
