+++ studyAi/src/app/library/question/components/server/answerComponent.tsx
@@ -1,34 +1,51 @@
 "use client";
-import { useQuery } from "@apollo/client";
-import { Question } from "../../../../../../prisma/generated/type-graphql";
-import { gql } from "@apollo/client";
 import { useParams } from "next/navigation";
 import { useQuestions } from "@/app/stores/questionStore";
-const getAnswerById = gql`
-  query Question($id: String) {
-    question(where: { id: $id }) {
-      id
-      answer {
-        answer
-      }
-    }
+import ContainerBar, { Container } from "./containerBar";
+import { IconButton } from "@mui/material";
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
+import {
+  faRefresh,
+  faUpRightAndDownLeftFromCenter,
+} from "@fortawesome/free-solid-svg-icons";
+import { QuestionTypes } from "@/app/util/types/UserData";
+const determineAnswerTitle = (str?: string) => {
+  const matchStr = str as (typeof QuestionTypes)[number];
+  switch (matchStr) {
+    case "multipleChoice":
+      return "Select the best answer";
+    case "selectMultiple":
+      return "Select all that apply";
+    case "shortAnswer":
+      return "Type your answer below";
+    default:
+      return str;
   }
-`;
-const AnswerContainer = ({ height }: { height?: string | number }) => {
+};
+const TopBar = () => {
   const params = useParams();
   const questions = useQuestions()[0].data;
-  const questionId = params?.id as string | undefined;
-  if (!questionId) return <></>;
-  const {
-    loading,
-    error,
-    data: queryData,
-  } = useQuery(getAnswerById, {
-    variables: { id: questionId },
-  });
-  const answerData = queryData as undefined | { question: Partial<Question> | null };
-  const question = questions[questionId];
-  console.log(loading, question, answerData);
-  return <></>;
+  const question =
+    params.id && typeof params.id === "string" ? questions[params.id] : null;
+  return (
+    <ContainerBar>
+      <h3>{determineAnswerTitle(question?.questionType)}</h3>
+      <div className="flex items-center h-full">
+        <IconButton>
+          <FontAwesomeIcon icon={faRefresh} />
+        </IconButton>
+        <IconButton>
+          <FontAwesomeIcon icon={faUpRightAndDownLeftFromCenter} />
+        </IconButton>
+      </div>
+    </ContainerBar>
+  );
+};
+const AnswerContainer = ({ height }: { height?: string | number }) => {
+  return (
+    <Container style={{ height: height ? height : undefined }}>
+      <TopBar />
+    </Container>
+  );
 };
 export default AnswerContainer;
