+++ studyAi/src/app/util/types/UserData.ts
@@ -1,3 +1,7 @@
 import { User } from "@prisma/client";
-
+export const QuestionTypes = [
+  "multipleChoice",
+  "selectMultiple",
+  "shortAnswer",
+] as const;
 export type UserInfo = User
